+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,29 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### 2.3.10-alpha.1 (2021-03-29)
+### 2.3.10-alpha.12 (2021-04-07)
 
 **Note:** Version bump only for package react-magma-dom
 
+### 2.3.10-alpha.11 (2021-04-07)
 
+**Note:** Version bump only for package react-magma-dom
 
+### 2.3.10-alpha.10 (2021-04-06)
 
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.9 (2021-04-06)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.8 (2021-04-05)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.7 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.6 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.1 (2021-04-01)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.5 (2021-03-31)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.4 (2021-03-31)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.3 (2021-03-30)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.2 (2021-03-30)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.1 (2021-03-29)
+
+**Note:** Version bump only for package react-magma-dom
 
 ### 2.3.10-alpha.0 (2021-03-25)
 
 **Note:** Version bump only for package react-magma-dom
 
-
-
-
-
 ### 2.3.9 (2021-03-24)
 
 **Note:** Version bump only for package react-magma-dom
-
-
-
-
 
 ### 2.3.9-alpha.13 (2021-03-23)